+++ pyrisco/__init__.py
@@ -1,1 +1,2 @@
-from .risco import RiscoAPI, CannotConnectError, OperationError, UnauthorizedError
+from .risco import get_risco_cloud, get_risco_local
+from .common import CannotConnectError, OperationError, UnauthorizedError
